code/requirements.txt
beautifulsoup4
pdfminer
regex
sklearn
pandas
xmltodict
matplotlib
python-Levenshtein-wheels